@jjsjann123 2025-01-22
phi3 now has a 450 us vs 494 us(ref). Which looks strange. I'll pull a profile to investigate.