@jjsjann123 2025-01-23
Two separate issues I'm seeing with benchmark number:
1. our profiler seems to be non-deterministically dropping cuda events. That's why running phi3 along has a longer measured kernel time;
2. standalone benchmark has faster kernel time on one of the segments.

Investigating.