@naoyam 2025-01-23
Yes, I'm looking at Phi3 but still only focuses on the forward fusion. 

I wonder if it's specific to Phi3? Or, could it be it was just because it was executed last? For example, if Qwen2 was executed after Phi3, would Qwen2 see a similar discrepancy?