@jjsjann123 2025-01-23
strangely that seems to only affect phi3.

I swapped it to run phi3 before qwen2 and haven't noticed any event dropped in qwen2. But given that it's not deterministic in phi3 in the first place, I'm not sure if this means there's anything specific to phi3.