@jjsjann123 2025-01-23
Re: missing event.

1. It's missing one of the small kernel at the beginning.
2. The first round seems to catch all 6 event, while the subsequent round were missing the very first kernel in the group.

So I'm now wondering if we can actually run

```
prof.start()
# measure
prof.stop()

prof.start()
# measure
prof.stop()
```

Patterns I see on pytorch are all using context manager, not sure if we are hitting a bug on profiler?