@Priya2698 2025-01-23
> I don't think this is specific to phi3. At least I'm seeing the same thing happening to qwen2, when it's not running as the first benchmark.
> 
> So there's indeed something wrong with the benchmark profiler and pytest.

Can you open an issue for this mentioning the order of benchmarks to reproduce this?
I'm wondering if the `start()`, `stop()` API was changed/deprecated.